# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.14.5'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /commands:
    get:
      operationId: getCommands
      summary: Fetch a list of Saved Commands
      tags:
        - Commands
      description: Without params, it returns a list of Saved Commands the user has access to
      parameters:
        - name: all
          in: query
          description: Can only be used by admins or managers to fetch all entities
          schema:
            type: boolean
        - name: userId
          in: query
          description: Standard users can use this only with their own _userId_
          schema:
            type: integer
        - name: deviceId
          in: query
          description: Standard users can use this only with _deviceId_s, they have access to
          schema:
            type: integer
        - name: groupId
          in: query
          description: >-
            Standard users can use this only with _groupId_s, they have access
            to
          schema:
            type: integer
        - name: refresh
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          description: Limit the number of returned results
          schema:
            type: integer
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
        - name: keyword
          in: query
          description: Search keyword filter
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Command'
        '401':
          description: Unauthorized
          content: {}
    post:
      operationId: postCommands
      summary: Create a Saved Command
      tags:
        - Commands
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Command'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
  /commands/{id}:
    get:
      operationId: getCommandsId
      summary: Fetch a Saved Command
      tags:
        - Commands
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not found
          content: {}
    put:
      operationId: putCommandsId
      summary: Update a Saved Command
      tags:
        - Commands
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Command'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
    delete:
      operationId: deleteCommandsId
      summary: Delete a Saved Command
      tags:
        - Commands
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
  /commands/send:
    get:
      operationId: getCommandsSend
      summary: Fetch a list of Saved Commands supported by Device at the moment
      description: >-
        Return a list of saved commands linked to Device and its groups,
        filtered by current Device protocol support
      tags:
        - Commands
      parameters:
        - name: deviceId
          in: query
          description: >-
            Standard users can use this only with _deviceId_s, they have access
            to
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Command'
        '400':
          description: Could happen when the user doesn't have permission for the device
          content: {}
    post:
      operationId: postCommandsSend
      summary: Dispatch commands to device
      description: Dispatch a new command or Saved Command if _body.id_ set
      tags:
        - Commands
      parameters:
        - name: groupId
          in: query
          description: Send the command to all devices in the group
          schema:
            type: integer
        - name: confirmed
          in: query
          description: Required for system commands marked as critical
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Command'
        required: true
      responses:
        '200':
          description: Command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Command'
        '202':
          description: Command queued
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/QueuedCommand'
                  - type: array
                    items:
                      $ref: '#/components/schemas/QueuedCommand'
        '400':
          description: >-
            Could happen when the user doesn't have permission or an incorrect
            command _type_ for the device
          content: {}
        '409':
          description: Critical confirmation missing or vehicle reported as moving
          content: {}
  /commands/defaults/bootstrap:
    post:
      operationId: postCommandsDefaultsBootstrap
      summary: Create the initial system command catalog if it does not exist
      description: Administrator only. This operation never sends a command to a device.
      tags:
        - Commands
      responses:
        '200':
          description: Current default command definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Command'
        '401':
          description: Unauthorized
          content: {}
  /commands/defaults/preview:
    get:
      operationId: getCommandsDefaultsPreview
      summary: Preview idempotent assignment of default commands to existing users
      tags:
        - Commands
      responses:
        '200':
          description: Assignment summary without modifying permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: integer
                  created:
                    type: integer
                  existing:
                    type: integer
                  ignoredUsers:
                    type: integer
                  commands:
                    type: array
                    items:
                      type: object
                  failures:
                    type: array
                    items:
                      type: string
    post:
      operationId: postCommandsDefaultsPreview
      summary: Preview assignment for selected users or users linked to selected groups
      tags:
        - Commands
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: integer
                groupIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Filtered assignment preview
          content:
            application/json:
              schema:
                type: object
  /commands/defaults/apply:
    post:
      operationId: postCommandsDefaultsApply
      summary: Assign missing default command permissions to eligible existing users
      description: Administrator only. Repeated calls are idempotent and never send commands.
      tags:
        - Commands
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: integer
                groupIds:
                  type: array
                  items:
                    type: integer
        required: true
      responses:
        '200':
          description: Assignment result
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content: {}
  /commands/types:
    get:
      operationId: getCommandsTypes
      summary: >-
        Fetch a list of available Commands for the Device or all possible
        Commands if Device ommited
      tags:
        - Commands
      parameters:
        - name: deviceId
          in: query
          description: >-
            Internal device identifier. Only works if device has already
            reported some locations
          schema:
            type: integer
        - name: textChannel
          in: query
          description: >-
            When `true` return SMS commands. If not specified or `false` return
            data commands
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandType'
        '400':
          description: >-
            Could happen when trying to fetch from a device the user does not
            have permission
          content: {}
...
  /permissions:
    get:
      operationId: getPermissions
      summary: Fetch permission links between Objects
      description: >-
        Provide exactly two `*Id` query parameters matching the Permission body
        shape (e.g. `reportId` and `deviceId`). Use `0` on a side to mean
        "any", e.g. `?reportId=5&deviceId=0` lists all devices linked to
        report 5. At least one side must be non-zero.
      tags:
        - Permissions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '400':
          description: Bad request
          content: {}
    post:
      operationId: postPermissions
      summary: Link an Object to another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: No permission
          content: {}
    delete:
      operationId: deletePermissions
      summary: Unlink an Object from another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
...
components:
  schemas:
...
    Command:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique saved command identifier
        deviceId:
          type: integer
          format: int64
          description: Target device identifier when the command is bound to one device
        description:
          type: string
          description: User friendly label displayed in the UI
        type:
          type: string
          description: Command type as defined by the device protocol
        textChannel:
          type: boolean
          description: Whether to send the command using the SMS channel
        attributes:
          type: object
          description: Additional parameters required by the command type
          properties: {}
    QueuedCommand:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identifier of the queued command job
        deviceId:
          type: integer
          format: int64
          description: Device identifier the queued command will be delivered to
        type:
          type: string
          description: Command type that will be executed
        textChannel:
          type: boolean
          description: Indicates whether the queued command uses SMS delivery
        attributes:
          type: object
          description: Stored parameters for the queued command
          properties: {}
...
    Permission:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: User id, can be only first parameter
        deviceId:
          type: integer
          format: int64
          description: >-
            Device id, can be first parameter or second only in combination with
            userId
        groupId:
          type: integer
          format: int64
          description: >-
            Group id, can be first parameter or second only in combination with
            userId
        geofenceId:
          type: integer
          format: int64
          description: Geofence id, can be second parameter only
        notificationId:
          type: integer
          format: int64
          description: Notification id, can be second parameter only
        calendarId:
          type: integer
          format: int64
          description: >-
            Calendar id, can be second parameter only and only in combination
            with userId
        attributeId:
          type: integer
          format: int64
          description: Computed attribute id, can be second parameter only
        driverId:
          type: integer
          format: int64
          description: Driver id, can be second parameter only
        managedUserId:
          type: integer
          format: int64
          description: >-
            User id, can be second parameter only and only in combination with
            userId
        commandId:
          type: integer
          format: int64
          description: Saved command id, can be second parameter only
      description: >-
        This is a permission map that contain two object indexes. It is used to
        link/unlink objects. Order is important. Example: { deviceId:8,
        geofenceId: 16 }
    CommandType:
      type: object
      properties:
        type:
          type: string
          description: Command type identifier
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer